be/NutritionalRecipeBook/src/NutritionalRecipeBook.Application/Services/CategoryService.cs
using NutritionalRecipeBook.Application.Contracts;
using NutritionalRecipeBook.Domain.Entities;
using NutritionalRecipeBook.Domain.Results;
using NutritionalRecipeBook.Infrastructure.Contracts;

namespace NutritionalRecipeBook.Application.Services
{
    public class CategoryService : ICategoryService
    {
        private readonly IGenericRepository<Category> _categoryRepository;

        public CategoryService(IGenericRepository<Category> categoryRepository)
        {
            _categoryRepository = categoryRepository;
        }

        public async Task<Result<List<Category>>> GetAllAsync()
        {
            var categories = await _categoryRepository.GetAllAsync();
            return Result<List<Category>>.Success(categories);
        }
    }
}
